[symex] Move the SSA equation and its passes into goto-symex/equation/
Fourth of seven commits partitioning src/goto-symex/
(docs/roadmap/goto-symex-partition-plan.md §2). symex_target and
symex_target_equation own SSA_stepst; slice, symex_symmetry and features are
its ssa_step_algorithm passes.

Pure move — the only content edits are #include paths, including the quoted
path-qualified form in src/util/ssa/goto_expr_factory.h (§7 trap 1).
