+++ src/esbmc/bmc.cpp
@@ -31,9 +31,9 @@
 #include <goto-symex/trace/build_goto_trace.h>
 #include <goto-symex/symex_invariant.h>
 #include <goto-symex/trace/goto_trace.h>
-#include <goto-symex/features.h>
+#include <goto-symex/equation/features.h>
 #include <goto-symex/trace/sarif.h>
-#include <goto-symex/symex_symmetry.h>
+#include <goto-symex/equation/symex_symmetry.h>
 #include <goto-symex/trace/xml_goto_trace.h>
 #include <langapi/language_util.h>
 #include <langapi/languages.h>
