+++ src/esbmc/bmc.h
@@ -4,9 +4,9 @@
 #include <goto-programs/dead_store_advisory.h>
 #include <goto-programs/goto_coverage.h>
 #include <goto-programs/property_verdict.h>
-#include <goto-symex/slice.h>
+#include <goto-symex/equation/slice.h>
 #include <goto-symex/reachability_tree.h>
-#include <goto-symex/symex_target_equation.h>
+#include <goto-symex/equation/symex_target_equation.h>
 #include <goto-symex/witness/witnesses.h>
 #include <goto-symex/testgen/pytest.h>
 #include <goto-symex/testgen/ctest.h>
