+++ src/goto-symex/equation/features.h
@@ -2,7 +2,7 @@
 
 #include <unordered_set>
 #include <util/ssa/algorithms.h>
-#include <goto-symex/symex_target_equation.h>
+#include <goto-symex/equation/symex_target_equation.h>
 
 enum class SSA_FEATURES
 {
