+++ src/goto-symex/equation/slice.h
@@ -1,7 +1,7 @@
 #ifndef CPROVER_GOTO_SYMEX_SLICE_H
 #define CPROVER_GOTO_SYMEX_SLICE_H
 
-#include <goto-symex/symex_target_equation.h>
+#include <goto-symex/equation/symex_target_equation.h>
 #include <util/base/time_stopping.h>
 #include <util/ssa/algorithms.h>
 #include <util/config/options.h>
