+++ src/goto-symex/equation/symex_target_equation.cpp
@@ -4,7 +4,7 @@
 #include <goto-symex/goto_symex.h>
 #include <goto-symex/goto_symex_state.h>
 #include <goto-symex/symex_invariant.h>
-#include <goto-symex/symex_target_equation.h>
+#include <goto-symex/equation/symex_target_equation.h>
 #include <langapi/language_util.h>
 #include <solvers/smt/smt_conv.h>
 #include <util/expr/expr_util.h>
