+++ src/goto-symex/goto_symex.h
@@ -5,7 +5,7 @@
 #include <util/base/threeval.h>
 #include <goto-programs/abstract-interpretation/interval_domain.h>
 #include <goto-symex/goto_symex_state.h>
-#include <goto-symex/symex_target.h>
+#include <goto-symex/equation/symex_target.h>
 #include <map>
 #include <optional>
 #include <pointer-analysis/dereference.h>
