+++ src/goto-symex/testgen/ctest.cpp
@@ -1,5 +1,5 @@
 #include <goto-symex/testgen/ctest.h>
-#include <goto-symex/slice.h>
+#include <goto-symex/equation/slice.h>
 #include <goto-symex/testgen/test_gen_guard.h>
 #include <goto-symex/witness/witnesses.h>
 #include <ac_config.h>
