+++ src/util/ssa/algorithms.h
@@ -4,7 +4,7 @@
 #include <goto-programs/goto_functions.h>
 #include <goto-programs/loopst.h>
 #include <goto-programs/goto_loops.h>
-#include <goto-symex/symex_target_equation.h>
+#include <goto-symex/equation/symex_target_equation.h>
 #include <util/message/message.h>
 /**
  * @brief Base interface to run an algorithm in esbmc
